# Speedup audb.available() for S3/Minio