--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -185,11 +185,14 @@ def persistent_repository(tmpdir_factory):
 
 @pytest.fixture(scope="module", autouse=False)
 def private_and_public_repository():
-    r"""Private and public repository on Artifactory.
+    r"""Private and public repositories.
 
     Configure the following repositories:
+
     * data-private: repo on public Artifactory without access
+    * audb-private: repo on public S3 without access
     * data-public: repo on public Artifactory with anonymous access
+    * audb-public: repo on public S3 with anonymous access
     * data-public2: repo on public Artifactory with anonymous access
 
     Note, that the order of the repos is important.
@@ -197,13 +200,15 @@ def private_and_public_repository():
     until it finds the requested database.
 
     """
-    host = "https://audeering.jfrog.io/artifactory"
-    backend = "artifactory"
     current_repositories = audb.config.REPOSITORIES
+    public_artifactory_host = "https://audeering.jfrog.io/artifactory"
+    public_s3_host = "s3.dualstack.eu-north-1.amazonaws.com"
     audb.config.REPOSITORIES = [
-        audb.Repository("data-private", host, backend),
-        audb.Repository("data-public", host, backend),
-        audb.Repository("data-public2", host, backend),
+        audb.Repository("data-private", public_artifactory_host, "artifactory"),
+        audb.Repository("audb-private", public_s3_host, "s3"),
+        audb.Repository("data-public", public_artifactory_host, "artifactory"),
+        audb.Repository("audb-public", public_s3_host, "s3"),
+        audb.Repository("data-public2", public_artifactory_host, "artifactory"),
     ]
 
     yield repository