This speeds up `audb.available()` for the S3 and MinIO backends by avoiding listing all files recursively, but focusing on the database names and the versions of the database header files.

Execution time for running `audb.available()` for the repositories `audb-public` (containing 7 smaller datasets) and `audb-internal` (containing 1 big dataset).

| Branch | audb-public | audb-internal |
| --- | --- | --- |
| speedup-available-s3 | 1.6 s | 0.4 s |
| main | 1.7 s | 45.0 s |

<details><summary> Benchmark code </summary>

```python
import audb
import time


repository = audb.Repository("audb-public", "s3.dualstack.eu-north-1.amazonaws.com", "s3")
audb.config.REPOSITORIES = [repository]
t0 = time.time()
df = audb.available()
t = time.time() - t0
print(f"Execution time: {t:.1f}s")
```

<details>

The pull request also makes sure `audb.available()` is tested for two S3 repositories as well, besides the Artifactory repositories.


## Summary by Sourcery

Enhancements:
- Optimize the audb.available() function for S3 and MinIO backends by avoiding recursive file listing and focusing on database names and header file versions.

## Summary by Sourcery

Enhancements:
- Optimize the audb.available() function for S3 and MinIO backends by avoiding recursive file listing and focusing on database names and header file versions.