@hagenw 2024-11-15
> Hey @hagenw - I've reviewed your changes - here's some feedback:
> 
> **Overall Comments**:
> 
> * Consider adding proper interface methods to the backend class instead of accessing _client directly. While the current approach works, it would be cleaner to maintain the abstraction layer. This could be addressed in a follow-up PR.

My original plan was also to add this feature directly to `audbackend`, by adding an argument `recursive` to `backend.ls()`. But the problem is, that `audbackend` does also support backends without any folder structure, for which `recursive` is meaningless. We could still discuss if it would make sense to add it to `audbackend`, but for now I would stay with the workaround of implementing it directly to `audb`. I created https://github.com/audeering/audbackend/issues/252 to track this.